# perf(sql): speed up GROUP BY queries with small fixed-size keys