--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -436,6 +436,8 @@ public interface CairoConfiguration {
 
     int getSqlSortValuePageSize();
 
+    int getSqlUnorderedMapMaxEntrySize();
+
     int getSqlWindowInitialRangeBufferSize();
 
     int getSqlWindowMaxRecursion();