--- a/core/src/main/java/io/questdb/cairo/map/OrderedMapRecord.java
+++ b/core/src/main/java/io/questdb/cairo/map/OrderedMapRecord.java
@@ -24,8 +24,9 @@
 
 package io.questdb.cairo.map;
 
-interface FastMapRecord extends MapRecord {
-    FastMapRecord clone();
+interface OrderedMapRecord extends MapRecord {
+
+    OrderedMapRecord clone();
 
     void of(long address);