--- a/core/src/main/java/io/questdb/cairo/map/Unordered16Map.java
+++ b/core/src/main/java/io/questdb/cairo/map/Unordered16Map.java
@@ -224,6 +224,7 @@ public class Unordered16Map implements Map, Reopenable {
             return;
         }
         Unordered16Map src16Map = (Unordered16Map) srcMap;
+        setKeyCapacity(size + src16Map.size);
 
         // First, we handle zero key.
         if (src16Map.hasZero && hasZero) {
@@ -262,10 +263,9 @@ public class Unordered16Map implements Map, Reopenable {
                 destAddr = getNextAddress(destAddr);
             }
 
+            assert free > 0;
             Vect.memcpy(destAddr, srcAddr, entrySize);
-            if (--free == 0) {
-                rehash();
-            }
+            free--;
             size++;
         }
     }