Includes the following:
* Introduces new hash tables: `Unordered8Map` and `Unordered16Map`. They're used in GROUP BY for small fixed-size keys (<= 8 and <= 16 bytes respectively) and small enough values. Unlike `FastMap`, these tables don't use separate heap which is beneficial performance-wise. Zero key value is used as the empty entry marker, so the zero key is kept in a separate chunk of memory.
* Renames `FastMap` to `OrderedMap`.
* Changes initial capacity for maps to 32. The previous default led to a lot of redundant space in the hash table in case of only a few keys.

```sql
-- FastMap: 2.9s
-- Unordered8Map: 1.7s
SELECT count_distinct(UserID) FROM hits;

-- FastMap: 770ms
-- Unordered8Map: 440ms
-- Rosti: 360ms
SELECT RegionID, SUM(AdvEngineID), COUNT(*) AS c, AVG(ResolutionWidth) FROM hits GROUP BY RegionID ORDER BY c DESC LIMIT 10;
```

Microbenchmarks:
```
Benchmark                               Mode  Cnt  Score    Error  Units
MapReadLongBenchmark.baseline           avgt    3  0.005 ±  0.001  us/op
MapReadLongBenchmark.testFastMap        avgt    3  0.151 ±  0.007  us/op
MapReadLongBenchmark.testHashMap        avgt    3  0.086 ±  0.006  us/op
MapReadLongBenchmark.testUnordered8Map  avgt    3  0.052 ±  0.002  us/op

Benchmark                                Mode  Cnt  Score    Error  Units
MapWriteLongBenchmark.baseline           avgt    3  0.007 ±  0.001  us/op
MapWriteLongBenchmark.testFastMap        avgt    3  0.122 ±  0.008  us/op
MapWriteLongBenchmark.testHashMap        avgt    3  0.151 ±  0.030  us/op
MapWriteLongBenchmark.testUnordered8Map  avgt    3  0.051 ±  0.012  us/op

Benchmark                                   Mode  Cnt  Score   Error  Units
MapReadLong128Benchmark.baseline            avgt    3  0.005 ± 0.001  us/op
MapReadLong128Benchmark.testFastMap         avgt    3  0.159 ± 0.014  us/op
MapReadLong128Benchmark.testUnordered16Map  avgt    3  0.053 ± 0.002  us/op

Benchmark                                    Mode  Cnt  Score   Error  Units
MapWriteLong128Benchmark.baseline            avgt    3  0.007 ± 0.002  us/op
MapWriteLong128Benchmark.testFastMap         avgt    3  0.132 ± 0.015  us/op
MapWriteLong128Benchmark.testUnordered16Map  avgt    3  0.056 ± 0.007  us/op
```