@ideoma 2024-01-14
#### [PR Coverage check]

:heart_eyes: **pass** : 643 / 721 (89.18%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMapValue.java|2|3|66.67%|
|:large_blue_circle:|io/questdb/cairo/map/UnorderedMapValue.java|85|113|75.22%|
|:large_blue_circle:|io/questdb/cairo/map/UnorderedMapRecord.java|99|120|82.50%|
|:large_blue_circle:|io/questdb/cairo/map/Unordered8Map.java|262|289|90.66%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMap.java|58|59|98.31%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/DistinctTimeSeriesRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/SampleByFillNoneRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/AsOfJoinRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinFilteredRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/window/RankFunctionFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/union/ExceptRecordCursorFactory.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/union/ExceptAllRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/LatestByRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/SampleByInterpolateRecordCursorFactory.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/window/RowNumberFunctionFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMapFixedSizeRecord.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/std/Hash.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/LatestByAllFilteredRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/DistinctRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/window/AvgDoubleWindowFunctionFactory.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/LatestByAllSymbolsFilteredRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinLightRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMapCursor.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/union/IntersectAllRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/union/UnionRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByAtom.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/LatestByLightRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/JoinRecordMetadata.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/window/SumDoubleWindowFunctionFactory.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/union/IntersectRecordCursorFactory.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/std/MemoryTag.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/window/FirstValueDoubleWindowFunctionFactory.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/LtJoinRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/SpliceJoinLightRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashJoinRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/OrderedMapVarSizeRecord.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/AbstractSampleByFillRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinFilteredLightRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashOuterJoinRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/HashJoinLightRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/UnorderedMapCursor.java|42|42|100.00%|
|:large_blue_circle:|io/questdb/cairo/map/MapFactory.java|23|23|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/LtJoinLightRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/join/AsOfJoinLightRecordCursorFactory.java|1|1|100.00%|